refactor: components
change the directory of components
